+++ pytest_tui/html_gen.py
@@ -12,14 +12,27 @@ from pytest_tui import __version__
 
 # from pytest_tui.__main__ import Cli
 # from pytest_tui.utils import CONFIGFILE, HTML_OUTPUT_FILE, TERMINAL_OUTPUT_FILE, Results
-from pytest_tui.utils import HTML_OUTPUT_FILE, TERMINAL_OUTPUT_FILE, Results
+from pytest_tui.utils import (
+    HTML_OUTPUT_FILE,
+    TERMINAL_OUTPUT_FILE,
+    Results,
+    test_session_starts_results_grabber,
+)
 
 CSS_FILE = Path(__file__).parent / "resources" / "styles.css"
 JS_FILE = Path(__file__).parent / "resources" / "scripts.js"
 
 TAB_METADATA = ["About"]
 TAB_METADATA_COLOR = {"About": "rgba(0, 0, 255, 0.33)"}
-TABS_RESULTS = ["All Tests", "Failures", "Passes", "Skipped", "Xfails", "Xpasses"]
+TABS_RESULTS = [
+    "All Tests",
+    "Failures",
+    "Passes",
+    "Skipped",
+    "Xfails",
+    "Xpasses",
+    "Reruns",
+]
 TABS_RESULTS_COLORS = {
     "All Tests": ("rgba(111, 104, 105, 0.34)", "rgba(128,128,0,0.5)"),
     "Failures": ("rgba(255, 10, 10, 0.75)", "rgba(255,0,0,0.5)"),
@@ -27,6 +40,7 @@ TABS_RESULTS_COLORS = {
     "Skipped": ("rgba(111, 104, 105, 0.34)", "rgba(128,128,0,0.5)"),
     "Xfails": ("rgba(255, 176, 176, 0.90)", "rgba(0,0,0,0.5)"),
     "Xpasses": ("rgba(0, 255, 0, 0.43)", "rgba(0,0,0,0.5)"),
+    "Reruns": ("rgba(220,165.0,1)", "rgba(255,194,0,0.5)"),
 }
 TABS_SECTIONS = [
     "summary_section",
@@ -34,6 +48,7 @@ TABS_SECTIONS = [
     "passes_section",
     "warnings_section",
     "errors_section",
+    "reruns_section",
 ]
 TABS_SECTIONS_COLORS = {
     "summary_section": "#daeaf6",
@@ -41,6 +56,7 @@ TABS_SECTIONS_COLORS = {
     "passes_section": "rgba(66, 228, 47, 0.6)",
     "warnings_section": "#ffee93",
     "errors_section": "#ffc09f",
+    "reruns_section": "#f6e3da",
 }
 TAB_FULL_OUTPUT = ["Console Output"]
 TAB_FULL_OUTPUT_COLOR = {"Console Output": "rgba(0, 0, 255, 0.33)"}
@@ -71,6 +87,7 @@ class TabContent:
         self.add("errors_section", self.results.tui_sections.errors.content)
         self.add("passes_section", self.results.tui_sections.passes.content)
         self.add("failures_section", self.results.tui_sections.failures.content)
+        self.add("reruns_section", self.results.tui_sections.rerun_test_summary.content)
 
         return self.get_all_items()
 
@@ -114,6 +131,9 @@ class HtmlPage:
             TABS_RESULTS.remove("Xpasses")
         if not self.results.tui_test_results.all_xfails():
             TABS_RESULTS.remove("Xfails")
+        # if len(self.results.tui_rerun_test_groups) == 0:
+        if not self.results.tui_test_results.all_reruns():
+            TABS_RESULTS.remove("Reruns")
 
     def create_header(self) -> str:
         my_css = Path(CSS_FILE).read_text().replace("\n", "")
@@ -128,13 +148,62 @@ class HtmlPage:
             + """</b></pre>"""
         )
 
+    def create_test_execution_info(self) -> str:
+        now = (
+            datetime.now(timezone.utc)
+            .replace(microsecond=0)
+            .strftime("%Y-%m-%d %H:%M:%S")
+        )
+        return (
+            """<button class="accordion">Test Execution Info</button><div class="panel"><p><pre>"""
+            + f"""<p>Test run started: {self.results.tui_session_start_time} UTC</p>"""
+            + f"""<p>Test run completed: {self.results.tui_session_end_time} UTC</p>"""
+            + f"""<p>Test run duration: {self.results.tui_session_duration}</p>"""
+            + f"""<p>This report generated: {now} UTC</p>"""
+            + f"""<p>pytest-tui version: {__version__}</p>"""
+            + """</pre></p></div>"""
+        )
+
+    def create_live_test_session_summary(self) -> str:
+        search = re.search(
+            test_session_starts_results_grabber,
+            self.results.tui_sections.test_session_starts.content.encode(
+                "unicode_escape"
+            ).decode(),
+        )
+        ripped = search.groups()[0].encode().decode("unicode-escape") if search else ""
+        return (
+            """<button class="accordion">Live Test Session Summary</button><div class="panel"><p><pre>"""
+            + self.converter.convert(
+                self.results.tui_sections.lastline.content, full=False
+            )
+            + self.converter.convert(ripped, full=False)
+            + self.converter.convert(
+                self.results.tui_sections.lastline.content, full=False
+            )
+            + """</pre></p></div>"""
+        )
+
     def create_testrun_summary(self) -> str:
         return (
-            """<h5>Short Summary:</h5><pre><b>"""
+            """<button class="accordion">Final Test Summary</button><div class="panel"><p><pre>"""
             + self.converter.convert(
                 self.results.tui_sections.short_test_summary.content, full=False
             )
-            + """</b></pre>"""
+            + "\n"
+            + self.converter.convert(
+                self.results.tui_sections.lastline.content, full=False
+            )
+            + """</pre></p></div>"""
+        )
+
+    def create_environment_info(self, m, table_attributes) -> str:
+        return (
+            """<button class="accordion">Environment</button><div class="panel"><p><pre>"""
+            + json2table.convert(
+                m, build_direction="LEFT_TO_RIGHT", table_attributes=table_attributes
+            )
+            + """</pre></p></div>"""
         )
 
     def create_trailer(self) -> str:
@@ -180,14 +249,18 @@ class HtmlPage:
         tab_links_section += """</div>"""
 
         tab_result_content = []
-        for result in TABS_RESULTS:
-            if result == "All Tests":
+        for tab in TABS_RESULTS:
+            if tab == "All Tests":
                 tab_result_content.append(
-                    f"""<div id="{result}" class="tabcontent"> {self.get_collapsible_results("tests")} </div>"""
+                    f"""<div id="{tab}" class="tabcontent"> {self.get_collapsible_results("tests")} </div>"""
+                )
+            elif tab == "Reruns":
+                tab_result_content.append(
+                    f"""<div id="{tab}" class="tabcontent"> {self.get_collapsible_results("reruns")} </div>"""
                 )
             else:
                 tab_result_content.append(
-                    f"""<div id="{result}" class="tabcontent"> {self.get_collapsible_results(result.lower())} </div>"""
+                    f"""<div id="{tab}" class="tabcontent"> {self.get_collapsible_results(tab.lower())} </div>"""
                 )
 
         tab_results = "".join(tab_result_content)
@@ -201,16 +274,27 @@ class HtmlPage:
 
         tab_metadata = f"""<div id="{TAB_METADATA[0]}" class="tabcontent"> <p>{self.get_metadata()}</p> </div>"""
 
-        tab_fullout = f"""<div id="{TAB_FULL_OUTPUT[0]}" class="tabcontent"> <pre>{self.get_terminal_output()}</pre> </div>"""
+        tab_full_output = f"""<div id="{TAB_FULL_OUTPUT[0]}" class="tabcontent"> <pre>{self.get_terminal_output()}</pre> </div>"""
 
         return (
-            tab_links_section + tab_metadata + tab_results + tab_sections + tab_fullout
+            tab_links_section
+            + tab_metadata
+            + tab_results
+            + tab_sections
+            + tab_full_output
         )
 
     def get_collapsible_results(self, outcome) -> str:
         collapsible_results = ""
-        results_by_outcome = eval(f"self.results.tui_test_results.all_{outcome}()")
-        for result in results_by_outcome:
+        if outcome == "reruns":
+            results = []
+            for group in self.results.tui_rerun_test_groups:
+                for result in group.full_test_list:
+                    results.append(result)
+        else:
+            results = eval(f"self.results.tui_test_results.all_{outcome}()")
+            print()
+        for result in results:
             content = self.converter.convert(
                 result.caplog
                 + result.capstderr
@@ -231,16 +315,6 @@ class HtmlPage:
             return ""
         m = json.loads(md.replace("'", '"').lstrip("metadata: "))
         m.pop("JAVA_HOME")
-        now = (
-            datetime.now(timezone.utc)
-            .replace(microsecond=0)
-            .strftime("%Y-%m-%d %H:%M:%S")
-        )
-        results_modification_time = (
-            datetime.utcfromtimestamp(Path(TERMINAL_OUTPUT_FILE).stat().st_mtime)
-            .replace(microsecond=0)
-            .strftime("%Y-%m-%d %H:%M:%S")
-        )
         table_attributes = {
             "id": "metadata",
             "font-family": "Helvetica, Arial, sans-serif",
@@ -253,15 +327,12 @@ class HtmlPage:
         }
 
         return (
-            f"{self.create_testrun_results()}<hr>"
+            # f"{self.create_testrun_results()}<hr>"
+            "<hr>"
+            + f"{self.create_test_execution_info()}<hr>"
+            + f"{self.create_live_test_session_summary()}<hr>"
             + f"{self.create_testrun_summary()}<hr>"
-            + f"""<h5>Test results generated:</h5> <p>{results_modification_time}</p>"""
-            + f"""<h5>This report generated:</h5> <p>{now}</p>"""
-            + f"""<h5>pytest-tui version:</h5> <p>{__version__}</p><hr>"""
-            + """<h5><b>Meta Data / Environment:</b></h5>"""
-            + json2table.convert(
-                m, build_direction="LEFT_TO_RIGHT", table_attributes=table_attributes
-            )
+            + f"{self.create_environment_info(m, table_attributes)}"
         )
 
     def get_terminal_output(self) -> str:
@@ -288,7 +359,7 @@ def main():
     with open(HTML_OUTPUT_FILE, "w+") as f:
         f.write(html_out)
 
-     # Open in browser
+    # Open in browser
     webbrowser.open(f"file://{HTML_OUTPUT_FILE._str}")
 
     # Open in browser if autolaunch_html config is set